+++ frontend/public/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- Added Bulk Actions to the Saved Sets Page. You can now, force recheck of autodownload, apply tags/labels and delete sets.
 - Added total number of sets for each MediUX User in the search results for better context.
 - Changed default tab on User Page to "Show Sets" and "Movie Sets" when no Box Sets are available.
 - Changed library filter option on User Page to not require unselecting of previous library selection.